LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-window/submit-samedocument-crossorigin-sameorigindomain.sub-expected.txt



FAIL submitting a <form> element that navigates cross-document targeting a same-origin-domain (but cross-origin) window assert_equals: navigationType expected "push" but got "replace"
